move package games into content
